Update custom calendar component.
